// Frontend/src/api/generated/model/trainerUserRelationship.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */

/**
 * Represents a relationship between a trainer and a user
 */
export interface TrainerUserRelationship {
  /** The user's unique identifier */
  user_id: string
  /** The trainer's unique identifier */
  trainer_id: string
  /**
   * The state of the relationship
   * @pattern ^(pending|active)$
   */
  state?: string
  /** When the relationship was created */
  created_at?: string
  /** When the relationship was last updated */
  updated_at?: string
}